Crate srml_contracts

Contract Module

The Contract module provides functionality for the runtime to deploy and execute WebAssembly smart-contracts.

contract::Trait

Call

Overview

This module extends accounts based on the Currency trait to have smart-contract functionality. It can be used with other modules that implement accounts based on Currency. These "smart-contract accounts" have the ability to create smart-contracts and make calls to other contract and non-contract accounts.

The smart-contract code is stored once in a code_cache, and later retrievable via its code_hash. This means that multiple smart-contracts can be instantiated from the same code_cache, without replicating the code each time.

When a smart-contract is called, its associated code is retrieved via the code hash and gets executed. This call can alter the storage entries of the smart-contract account, create new smart-contracts, or call other smart-contracts.

Finally, when an account is reaped, its associated code and storage of the smart-contract account will also be deleted.

Gas

Senders must specify a gas limit with every call, as all instructions invoked by the smart-contract require gas. Unused gas is refunded after the call, regardless of the execution outcome.

If the gas limit is reached, then all calls and state changes (including balance transfers) are only reverted at the current call's contract level. For example, if contract A calls B and B runs out of gas mid-call, then all of B's calls are reverted. Assuming correct error handling by contract A, A's other calls and state changes still persist.

Notable Scenarios

Contract call failures are not always cascading. When failures occur in a sub-call, they do not "bubble up", and the call will only revert at the specific contract level. For example, if contract A calls contract B, and B fails, A can decide how to handle that failure, either proceeding or reverting A's changes.

Interface

Dispatchable functions

put_code - Stores the given binary Wasm code into the chain's storage and returns its code_hash.

create - Deploys a new contract from the given code_hash, optionally transferring some balance. This creates a new smart contract account and calls its contract deploy handler to initialize the contract.

call - Makes a call to an account, optionally transferring some balance.

Usage

The Contract module is a work in progress. The following examples show how this Contract module can be used to create and call contracts.

ink is an eDSL that enables writing WebAssembly based smart contracts in the Rust programming language. This is a work in progress.

Related Modules

Balances

Macros

convert_args

define_func

gen_signature

gen_signature_dispatch

match_tokens
A simple utility macro that helps to match against a list of tokens.

register_func

unmarshall_then_body_then_marshall

Structs

AccountCounter
The subtrie counter.

CodeStorage
A mapping between an original code hash and instrumented wasm code, ready for execution.

Config
In-memory cache of configuration values.

ContractInfoOf
The code associated with a given account.

DefaultBlockGasLimit
A resonable default value for Trait::BlockGasLimit.

DefaultCallBaseFee
A resonable default value for Trait::CallBaseFee.

DefaultContractFee
A resonable default value for Trait::ContractFee.

DefaultCreateBaseFee
A resonable default value for Trait::CreateBaseFee.

DefaultCreationFee
A resonable default value for Trait::CreationFee.

DefaultDispatchFeeComputor
The default dispatch fee computor computes the fee in the same way that the implementation of MakePayment for the Balances module does.

DefaultMaxDepth
A resonable default value for Trait::MaxDepth.

DefaultRentByteFee
A resonable default value for Trait::RentByteFee.

DefaultRentDepositOffset
A resonable default value for Trait::RentDepositOffset.

DefaultSignedClaimHandicap
A resonable default value for [Trait::SignedClaimedHandicap].

DefaultStorageSizeOffset
A resonable default value for Trait::StorageSizeOffset.

DefaultSurchargeReward
A resonable default value for Trait::SurchargeReward.

DefaultTombstoneDeposit
A resonable default value for Trait::TombstoneDeposit.

DefaultTransactionBaseFee
A resonable default value for Trait::TransactionBaseFee.

DefaultTransactionByteFee
A resonable default value for Trait::TransactionByteFee.

DefaultTransferFee
A resonable default value for Trait::TransferFee.

GenesisConfig

Module

PristineCode
A mapping from an original code hash to the original code, untouched by instrumentation.

RawAliveContractInfo
Information for managing an account and its sub trie abstraction. This is the required info to cache for an account.

RawTombstoneContractInfo

Schedule
Definition of the cost schedule and other parameterizations for wasm vm.

SimpleAddressDeterminator
Simple contract address determiner.

TrieIdFromParentCounter
Get trie id from account_id.

Enums

Call
Contracts module.

ContractInfo
Information for managing an acocunt and its sub trie abstraction. This is the required info to cache for an account

RawEvent
Events for this module.

Traits

ComputeDispatchFee
A function that returns the fee for dispatching a Call.

ContractAddressFor
A function that generates an AccountId for a contract upon instantiation.

Trait

TrieIdGenerator
Get a trie id (trie id must be unique and collision resistant depending upon its context). Note that it is different than encode because trie id should be collision resistant (being a proper unique identifier).

Type Definitions

AliveContractInfo

BalanceOf

CodeHash

Event
RawEvent specialized for the configuration Trait

Gas

NegativeImbalanceOf

TombstoneContractInfo

TrieId