Trait srml_contracts::Trait

pub trait Trait: Trait {
    type Currency: Currency<Self::AccountId>;
    type Time: Time;
    type Randomness: Randomness<Self::Hash>;
    type Call: Parameter + Dispatchable<Origin = Self::Origin> + IsSubType<Module<Self>, Self>;
    type Event: From<Event<Self>> + Into<Self::Event>;
    type DetermineContractAddress: ContractAddressFor<CodeHash<Self>, Self::AccountId>;
    type ComputeDispatchFee: ComputeDispatchFee<Self::Call, BalanceOf<Self>>;
    type TrieIdGenerator: TrieIdGenerator<Self::AccountId>;
    type GasPayment: OnUnbalanced<NegativeImbalanceOf<Self>>;
    type RentPayment: OnUnbalanced<NegativeImbalanceOf<Self>>;
    type SignedClaimHandicap: Get<Self::BlockNumber>;
    type TombstoneDeposit: Get<BalanceOf<Self>>;
    type StorageSizeOffset: Get<u32>;
    type RentByteFee: Get<BalanceOf<Self>>;
    type RentDepositOffset: Get<BalanceOf<Self>>;
    type SurchargeReward: Get<BalanceOf<Self>>;
    type TransferFee: Get<BalanceOf<Self>>;
    type CreationFee: Get<BalanceOf<Self>>;
    type TransactionBaseFee: Get<BalanceOf<Self>>;
    type TransactionByteFee: Get<BalanceOf<Self>>;
    type ContractFee: Get<BalanceOf<Self>>;
    type CallBaseFee: Get<Gas>;
    type InstantiateBaseFee: Get<Gas>;
    type MaxDepth: Get<u32>;
    type MaxValueSize: Get<u32>;
    type BlockGasLimit: Get<Gas>;
}

Associated Types

type Currency: Currency<Self::AccountId>

type Time: Time

type Randomness: Randomness<Self::Hash>

type Call: Parameter + Dispatchable<Origin = Self::Origin> + IsSubType<Module<Self>, Self>

The outer call dispatch type.

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type DetermineContractAddress: ContractAddressFor<CodeHash<Self>, Self::AccountId>

A function type to get the contract address given the instantiator.

type ComputeDispatchFee: ComputeDispatchFee<Self::Call, BalanceOf<Self>>

A function type that computes the fee for dispatching the given Call.

It is recommended (though not required) for this function to return a fee that would be taken by the Executive module for regular dispatch.

type TrieIdGenerator: TrieIdGenerator<Self::AccountId>

trie id generator

type GasPayment: OnUnbalanced<NegativeImbalanceOf<Self>>

Handler for the unbalanced reduction when making a gas payment.

type RentPayment: OnUnbalanced<NegativeImbalanceOf<Self>>

Handler for rent payments.

type SignedClaimHandicap: Get<Self::BlockNumber>

Number of block delay an extrinsic claim surcharge has.

When claim surcharge is called by an extrinsic the rent is checked for current_block - delay

type TombstoneDeposit: Get<BalanceOf<Self>>

The minimum amount required to generate a tombstone.

type StorageSizeOffset: Get<u32>

Size of a contract at the time of instantiation. This is a simple way to ensure that empty contracts eventually gets deleted.

type RentByteFee: Get<BalanceOf<Self>>

Price of a byte of storage per one block interval. Should be greater than 0.

type RentDepositOffset: Get<BalanceOf<Self>>

The amount of funds a contract should deposit in order to offset the cost of one byte.

Let's suppose the deposit is 1,000 BU (balance units)/byte and the rent is 1 BU/byte/day, then a contract with 1,000,000 BU that uses 1,000 bytes of storage would pay no rent. But if the balance reduced to 500,000 BU and the storage stayed the same at 1,000, then it would pay 500 BU/day.

type SurchargeReward: Get<BalanceOf<Self>>

Reward that is received by the party whose touch has led to removal of a contract.

type TransferFee: Get<BalanceOf<Self>>

The fee required to make a transfer.

type CreationFee: Get<BalanceOf<Self>>

The fee required to create an account.

type TransactionBaseFee: Get<BalanceOf<Self>>

The fee to be paid for making a transaction; the base.

type TransactionByteFee: Get<BalanceOf<Self>>

The fee to be paid for making a transaction; the per-byte portion.

type ContractFee: Get<BalanceOf<Self>>

The fee required to instantiate a contract instance.

type CallBaseFee: Get<Gas>

The base fee charged for calling into a contract.

type InstantiateBaseFee: Get<Gas>

The base fee charged for instantiating a contract.

type MaxDepth: Get<u32>

The maximum nesting level of a call/instantiate stack.

type MaxValueSize: Get<u32>

The maximum size of a storage value in bytes.

type BlockGasLimit: Get<Gas>

The maximum amount of gas that could be expended per block.

Loading content...