Trait srml_contracts::TrieIdGenerator

pub trait TrieIdGenerator<AccountId> {
    fn trie_id(account_id: &AccountId) -> TrieId;
}

Get a trie id (trie id must be unique and collision resistant depending upon its context). Note that it is different than encode because trie id should be collision resistant (being a proper unique identifier).

Required methods

fn trie_id(account_id: &AccountId) -> TrieId

Get a trie id for an account, using reference to parent account trie id to ensure uniqueness of trie id.

The implementation must ensure every new trie id is unique: two consecutive calls with the same parameter needs to return different trie id values.

Also, the implementation is responsible for ensuring that TrieId starts with :child_storage:. TODO: We want to change this, see https://github.com/paritytech/substrate/issues/2325

Loading content...